file cputime.c edited #6.

+++ cputime.c
@@ -36,7 +36,7 @@ int main(int argc, char *argv[])
    */
   prev = 1;
   succ = 2;
-  for(i = 1; i < (10 * rand()) / RAND_MAX; i++) {
+  for(i = 1; i < (10 * (int) ((float) rand() / (float) RAND_MAX)); i++) {
     printf("%d\n", succ);
     temp = prev + succ;
     prev = succ;
